requirements.txt
Flask
requests
python_dotenv
chatterbot
chatterbot_corpus